gpu: nvgpu: Fix gv11b_gr_set_sm_debug_mode for gpus with more than 32 sms
For gpu's 32 <  number of sms <= 64, the hal uses integer type (which is usually
32 bit) for checking masks and left shifts the integer with values greater
than 32.
To avoid this is undefined behaviour, use u64 instead.

Bug 2418354

Reviewed-on: https://git-master.nvidia.com/r/1926890
Tested-by: Thomas Fleury <tfleury@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/gv11b/gr_gv11b.c
@@ -3222,7 +3222,7 @@ int gv11b_gr_set_sm_debug_mode(struct gk20a *g,
 		u32 gpc, tpc, sm;
 		u32 reg_offset, reg_mask, reg_val;
 
-		if (!(sms & (1 << sm_id))) {
+		if (!(sms & ((u64)1 << sm_id))) {
 			continue;
 		}
